+++ packages/terra-i18n/tests/jest/i18nLoader.test.js
@@ -48,31 +48,26 @@ describe('i18nLoader', () => {
 
     it('loads en locale', () => {
       i18nLoader('en', jest.fn());
-    });
 
-    it('adds the intl polyfill', () => {
+      // adds the intl polyfill
       expect(global.Intl).toBeDefined();
       expect(global.IntlPolyfill).toBeDefined();
-    });
 
-    it('loads the locale data ', () => {
+      // loads the locale data
       expect(defaultLoadIntl).toHaveBeenNthCalledWith(1, 'en');
-    });
 
-    it('loads translations', () => {
+      // loads translations
       expect(loadTranslations).toHaveBeenNthCalledWith(1, 'en', expect.anything(), undefined);
     });
 
     it('loads es locale', () => {
       i18nLoader('es', jest.fn());
-    });
 
-    it('loads the new locale data on sequential load ', () => {
-      expect(defaultLoadIntl).toHaveBeenNthCalledWith(2, 'es');
-    });
+      // loads the new locale data on sequential load
+      expect(defaultLoadIntl).toHaveBeenNthCalledWith(1, 'es');
 
-    it('loads the new translation data on sequential load ', () => {
-      expect(loadTranslations).toHaveBeenNthCalledWith(2, 'es', expect.anything(), undefined);
+      // loads the new translation data on sequential load
+      expect(loadTranslations).toHaveBeenNthCalledWith(1, 'es', expect.anything(), undefined);
     });
   });
 
@@ -86,31 +81,26 @@ describe('i18nLoader', () => {
 
     it('loads en locale', () => {
       i18nLoader('en', jest.fn());
-    });
 
-    it('adds the intl polyfill', () => {
+      // adds the intl polyfill
       expect(global.Intl).toBeDefined();
       expect(global.IntlPolyfill).toBeDefined();
-    });
 
-    it('loads the locale data ', () => {
+      // loads the locale data
       expect(defaultLoadIntl).toHaveBeenNthCalledWith(1, 'en');
-    });
 
-    it('loads translations', () => {
+      // loads translations
       expect(loadTranslations).toHaveBeenNthCalledWith(1, 'en', expect.anything(), undefined);
     });
 
     it('loads es locale', () => {
       i18nLoader('es', jest.fn());
-    });
 
-    it('loads the new locale data on sequential load ', () => {
-      expect(defaultLoadIntl).toHaveBeenNthCalledWith(2, 'es');
-    });
+      // loads the new locale data on sequential load
+      expect(defaultLoadIntl).toHaveBeenNthCalledWith(1, 'es');
 
-    it('loads the new translation data on sequential load ', () => {
-      expect(loadTranslations).toHaveBeenNthCalledWith(2, 'es', expect.anything(), undefined);
+      // loads the new translation data on sequential load
+      expect(loadTranslations).toHaveBeenNthCalledWith(1, 'es', expect.anything(), undefined);
     });
   });
 
@@ -124,31 +114,26 @@ describe('i18nLoader', () => {
 
     it('loads en locale', () => {
       i18nLoader('en', jest.fn());
-    });
 
-    it('adds the intl polyfill', () => {
+      // adds the intl polyfill
       expect(global.Intl).toBeDefined();
       expect(global.IntlPolyfill).toBeDefined();
-    });
 
-    it('loads the locale data ', () => {
+      // loads the locale data
       expect(defaultLoadIntl).toHaveBeenNthCalledWith(1, 'en');
-    });
 
-    it('loads translations', () => {
+      // loads translations
       expect(loadTranslations).toHaveBeenNthCalledWith(1, 'en', expect.anything(), undefined);
     });
 
     it('loads es locale', () => {
       i18nLoader('es', jest.fn());
-    });
 
-    it('loads the new locale data on sequential load ', () => {
-      expect(defaultLoadIntl).toHaveBeenNthCalledWith(2, 'es');
-    });
+      // loads the new locale data on sequential load
+      expect(defaultLoadIntl).toHaveBeenNthCalledWith(1, 'es');
 
-    it('loads the new translation data on sequential load ', () => {
-      expect(loadTranslations).toHaveBeenNthCalledWith(2, 'es', expect.anything(), undefined);
+      // loads the new translation data on sequential load
+      expect(loadTranslations).toHaveBeenNthCalledWith(1, 'es', expect.anything(), undefined);
     });
   });
 });