+++ jest.globalSetup.js
@@ -1,0 +1,6 @@
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+const aggregateTranslations = require('@cerner/terra-aggregate-translations');
+
+module.exports = () => {
+  aggregateTranslations();
+};