+++ jestglobalsetup.js
@@ -1,6 +1,0 @@
-/* eslint-disable-next-line import/no-extraneous-dependencies */
-const aggregateTranslations = require('terra-aggregate-translations');
-
-module.exports = () => {
-  aggregateTranslations();
-};