+++ packages/terra-paginator/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Changed
   * Removed unnecessary `wdio.conf.js` file
   * Upgrade tests to terra-functional-testing and WebDriverIO v6
+  * Updated Jest command
 
 ## 2.70.0 - (March 2, 2021)